spirv: disable main()
Exporting a main function with no arguments makes no sense in a
shader. This does make spirv a backend where an executable does
not automatically reference main, but that is probably not
something that you want anyway. In this case, an executable
is just a fully linked program, in contrast to a library.

This also helps with testing infrastructure, as the main()
function from the test runner does not automatically get
exported this way. In the future this may need to be changed
(if function pointers are ever going to be supported), but
in that situation we can explicitly export the entry point
from the test runner.